rdp: Removed EACK
The EACKs are part of the RDP RFC. However, for satellite communication
the link may be degraded by EACKs because they are rapidly generated and
otherwise transmitted out ot sequence. (1 ack per packet)
This messes with the rdp_options and we have found this to be both
confusing and detrimental to reliable communication. Especially when
dealing with half-duplex links.
This commit simply stops generating EACKs and relies instead on the
normal packet timeout and retransmission procedure. It has been tested
for almost a year in production at Space Inventor.