cmake: Enable -Wpedantic flag
To ensure compatibility and avoid non-standard features like GNU
extensions in libraries such as libcsp, we enable the -Wpedantic
flag. This flag helps enforce adherence to standard C by issuing
warnings when extensions are used.

Since we rely on the token-pasting operator, which isn't standard in
C, Clang generates a warning for it. As this feature is necessary and
simplifies the macro, we suppress the warning in Clang with
-Wno-gnu-zero-variadic-macro-arguments.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>