experimental cygwin port
These changes need to be applied for libcsp to be compiled on fresh cygwin
installation on windows 11.

Ethernet packet forge API is not available with cygwin so it has been
disabled with a warning at configuration and compilation time.
If this API is needed under cygwin, a similar and portable API is supposed
to be provided by libpcap (check also https://pcapplusplus.github.io).

This port has been proven useful for an internal project and may benefit to
others.

The only build system which has been updated for cygwin is CMake.

Limitation to CMake, and Ethernet removal is the justification for
the "experimental" mention in the summary.